# docs/openapi/marcus-trading-delta.yaml
openapi: 3.0.3
info:
  title: Marcus Trading API Delta
  version: 0.2.0
  description: >-
    Contract-first delta for Marcus trading integration.
    This version completes High and Medium priority operations and standardizes
    reusable error responses.
  license:
    name: Proprietary
    url: https://marcus.local/license
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /signals:
    post:
      tags: [system]
      operationId: submitSignal
      summary: Ingest a trading signal into Marcus routing pipeline.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignalSubmissionRequest'
      responses:
        '200':
          description: Signal accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalSubmissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [system]
      operationId: listSignals
      summary: List recent signals.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, DISPATCHED, FAILED, ALL]
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Signals.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SignalItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /dashboard/overview:
    get:
      tags: [dashboard]
      operationId: getDashboardOverview
      summary: Load KPI cards and top-level trading overview.
      responses:
        '200':
          description: Dashboard KPIs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardOverviewResponse'
              examples:
                default:
                  value:
                    totalEquity: 12500.25
                    openPnl: 132.4
                    winRate: 0.61
                    activeBots: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /dashboard/equity-series:
    get:
      tags: [dashboard]
      operationId: getDashboardEquitySeries
      summary: Load equity curve by selected range.
      parameters:
        - $ref: '#/components/parameters/Range'
      responses:
        '200':
          description: Time series points.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSeriesPoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /dashboard/exchange-allocation:
    get:
      tags: [dashboard]
      operationId: getExchangeAllocation
      summary: Load exchange allocation for donut chart.
      responses:
        '200':
          description: Allocation list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExchangeAllocationItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /leaderboard/strategies:
    get:
      tags: [leaderboard]
      operationId: listLeaderboardStrategies
      summary: List ranked strategies for leaderboard tables.
      parameters:
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [24H, 7D, 30D, ALL]
        - name: market
          in: query
          schema:
            type: string
        - name: asset
          in: query
          schema:
            type: string
        - name: rankMetric
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Leaderboard page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardStrategiesPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /leaderboard/featured:
    get:
      tags: [leaderboard]
      operationId: listLeaderboardFeatured
      summary: Get featured top leaderboard cards.
      responses:
        '200':
          description: Featured list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardFeaturedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /leaderboard/spotlights:
    get:
      tags: [leaderboard]
      operationId: listStrategySpotlights
      summary: Get strategy spotlight cards.
      responses:
        '200':
          description: Spotlights.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StrategySpotlight'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /paper/session:
    get:
      tags: [paper]
      operationId: getPaperSessionSummary
      summary: Load paper trading session summary.
      responses:
        '200':
          description: Session summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperSessionSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /paper/signals:
    get:
      tags: [paper]
      operationId: listPaperSignals
      summary: List active or recent paper signals.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, EXECUTED, EXPIRED, ALL]
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Signals.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaperSignalItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /paper/orders:
    post:
      tags: [paper]
      operationId: createPaperOrder
      summary: Create paper order from quick order form.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaperOrderRequest'
      responses:
        '201':
          description: Order accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperOrderResponse'
              examples:
                default:
                  value:
                    orderId: ord_001
                    status: ACCEPTED
                    executedPrice: 64500.5
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /paper/session/pause:
    post:
      tags: [paper]
      operationId: pausePaperSession
      summary: Pause current paper trading session.
      responses:
        '200':
          description: Session paused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperSessionStateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /paper/session/resume:
    post:
      tags: [paper]
      operationId: resumePaperSession
      summary: Resume current paper trading session.
      responses:
        '200':
          description: Session resumed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperSessionStateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /paper/logs:
    get:
      tags: [paper]
      operationId: listPaperExecutionLogs
      summary: Paginated execution logs for paper trading.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Log page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperExecutionLogPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Range:
      name: range
      in: query
      required: true
      schema:
        type: string
        enum: [1D, 1W, 1M, YTD, ALL]
  responses:
    BadRequest:
      description: Bad request.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource conflict.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationFailed:
      description: Validation failed.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    InternalServerError:
      description: Internal server error.
      headers:
        X-Trace-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    OffsetPaginationMeta:
      type: object
      required: [page, size, totalElements, totalPages, hasNext]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean
    CursorPaginationMeta:
      type: object
      required: [cursor, nextCursor, limit, hasMore]
      properties:
        cursor:
          type: string
          nullable: true
        nextCursor:
          type: string
          nullable: true
        limit:
          type: integer
        hasMore:
          type: boolean
...
    SignalSubmissionRequest:
      type: object
      required: [signalId, botId, exchangeSlug, symbol, action, price, timestamp]
      properties:
        signalId:
          type: string
        botId:
          type: string
        exchangeSlug:
          type: string
        symbol:
          type: string
        action:
          type: string
          enum: [OPEN_LONG, OPEN_SHORT, CLOSE_LONG, CLOSE_SHORT, UPDATE_TP_SL]
        price:
          type: number
          format: double
        metadata:
          type: object
          additionalProperties:
            type: object
        timestamp:
          type: string
          format: date-time
    SignalSubmissionResponse:
      type: object
      properties:
        signalId:
          type: string
        status:
          type: string
        acceptedAt:
          type: string
          format: date-time
    SignalItem:
      type: object
      properties:
        signalId:
          type: string
        botId:
          type: string
        exchangeSlug:
          type: string
        symbol:
          type: string
        action:
          type: string
        price:
          type: number
          format: double
        status:
          type: string
        generatedTimestamp:
          type: string
          format: date-time
...
    DashboardOverviewResponse:
      type: object
      properties:
        totalEquity:
          type: number
          format: double
        openPnl:
          type: number
          format: double
        winRate:
          type: number
          format: double
        activeBots:
          type: integer
    TimeSeriesPoint:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
          format: double
    ExchangeAllocationItem:
      type: object
      properties:
        exchange:
          type: string
        percentage:
          type: number
          format: double
...
    LeaderboardStrategyItem:
      type: object
      properties:
        rank:
          type: integer
        strategyId:
          type: string
        strategyName:
          type: string
        creatorName:
          type: string
        cagr:
          type: number
          format: double
        sharpe:
          type: number
          format: double
        maxDrawdown:
          type: number
          format: double
    LeaderboardStrategiesPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LeaderboardStrategyItem'
        meta:
          $ref: '#/components/schemas/OffsetPaginationMeta'
    LeaderboardFeaturedItem:
      type: object
      properties:
        strategyId:
          type: string
        strategyName:
          type: string
        rankLabel:
          type: string
        sharpe:
          type: number
          format: double
    LeaderboardFeaturedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LeaderboardFeaturedItem'
    StrategySpotlight:
      type: object
      properties:
        strategyId:
          type: string
        strategyName:
          type: string
        market:
          type: string
        oneDayReturn:
          type: number
          format: double
    PaperSessionSummaryResponse:
      type: object
      properties:
        sessionId:
          type: string
        status:
          type: string
        virtualBalance:
          type: number
          format: double
        openPnl:
          type: number
          format: double
        buyingPower:
          type: number
          format: double
    PaperSignalItem:
      type: object
      properties:
        signalId:
          type: string
        botId:
          type: string
        assetPair:
          type: string
        side:
          type: string
        confidence:
          type: number
          format: double
        status:
          type: string
        generatedAt:
          type: string
          format: date-time
    CreatePaperOrderRequest:
      type: object
      required: [assetPair, orderType, side, quantity]
      properties:
        assetPair:
          type: string
        orderType:
          type: string
          enum: [MARKET, LIMIT]
        side:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: number
          format: double
        limitPrice:
          type: number
          format: double
    PaperOrderResponse:
      type: object
      properties:
        orderId:
          type: string
        status:
          type: string
        executedPrice:
          type: number
          format: double
    PaperSessionStateResponse:
      type: object
      properties:
        sessionId:
          type: string
        status:
          type: string
          enum: [RUNNING, PAUSED]
    PaperExecutionLogItem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: string
        message:
          type: string
    PaperExecutionLogPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaperExecutionLogItem'
        meta:
          $ref: '#/components/schemas/CursorPaginationMeta'
...
    ErrorResponse:
      type: object
      required: [code, message, status, traceId, path, timestamp]
      properties:
        code:
          type: string
          description: Stable application error code.
        message:
          type: string
        status:
          type: integer
          format: int32
        traceId:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
      example:
        code: BAD_REQUEST
        message: Bot name is required
        status: 400
        traceId: 7dc5e5b8-3136-4b71-8540-b6f85cb3b6b3
        path: /api/v1/bots
        timestamp: '2026-04-10T10:00:00Z'
    FieldValidationError:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              items:
                $ref: '#/components/schemas/FieldValidationError'
      example:
        code: VALIDATION_FAILED
        message: Validation failed
        status: 422
        traceId: 643ad58e-13a8-4f38-9bf1-6be7f0a093f0
        path: /api/v1/paper/orders
        timestamp: '2026-04-10T10:00:00Z'
        errors:
          - field: quantity
            reason: must be greater than 0